# docs/swagger/swagger.json
{
    "schemes": [
        "http",
        "https"
    ],
    "swagger": "2.0",
    "info": {
        "description": "REST API for JAV metadata scraping and file organization",
        "title": "Javinizer API",
        "termsOfService": "https://github.com/javinizer/javinizer-go",
        "contact": {
            "name": "API Support",
            "url": "https://github.com/javinizer/javinizer-go/issues"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/",
    "paths": {
...
        "/api/v1/history": {
            "get": {
                "description": "Get a paginated list of history records with optional filtering",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "history"
                ],
                "summary": "List history records",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Number of records to return (default: 50, max: 500)",
                        "name": "limit",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Number of records to skip (default: 0)",
                        "name": "offset",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Filter by operation type (scrape, organize, download, nfo)",
                        "name": "operation",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Filter by status (success, failed, reverted)",
                        "name": "status",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Filter by movie ID",
                        "name": "movie_id",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/internal_api_history.HistoryListResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/github_com_javinizer_javinizer-go_internal_api_contracts.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/github_com_javinizer_javinizer-go_internal_api_contracts.ErrorResponse"
                        }
                    }
                }
            },
            "delete": {
                "description": "Delete multiple history records based on criteria",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "history"
                ],
                "summary": "Delete history records in bulk",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Delete records older than N days",
                        "name": "older_than_days",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Delete all records for a specific movie",
                        "name": "movie_id",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/internal_api_history.DeleteHistoryBulkResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/github_com_javinizer_javinizer-go_internal_api_contracts.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/github_com_javinizer_javinizer-go_internal_api_contracts.ErrorResponse"
                        }
                    }
                }
            }
        },
...
        "/api/v1/jobs/{id}": {
            "get": {
                "description": "Get details for a specific batch job by ID",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "jobs"
                ],
                "summary": "Get a single batch job",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Job ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github_com_javinizer_javinizer-go_internal_api_contracts.JobListItem"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/github_com_javinizer_javinizer-go_internal_api_contracts.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/github_com_javinizer_javinizer-go_internal_api_contracts.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/api/v1/jobs/{id}/operations": {
            "get": {
                "description": "Get all file operations for a specific batch job with before/after paths",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "jobs"
                ],
                "summary": "List operations for a batch job",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Job ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github_com_javinizer_javinizer-go_internal_api_contracts.OperationListResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/github_com_javinizer_javinizer-go_internal_api_contracts.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/github_com_javinizer_javinizer-go_internal_api_contracts.ErrorResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "github_com_javinizer_javinizer-go_internal_api_contracts.ErrorResponse": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string",
                    "example": "Movie not found"
                },
                "errors": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                }
            }
        },
...
        "github_com_javinizer_javinizer-go_internal_api_contracts.JobListItem": {
            "type": "object",
            "properties": {
                "completed": {
                    "type": "integer",
                    "example": 9
                },
                "completed_at": {
                    "type": "string",
                    "example": "2026-04-12T10:05:00Z"
                },
                "destination": {
                    "type": "string",
                    "example": "/path/to/output"
                },
                "failed": {
                    "type": "integer",
                    "example": 1
                },
                "id": {
                    "type": "string",
                    "example": "550e8400-e29b-41d4-a716-446655440000"
                },
                "operation_count": {
                    "type": "integer",
                    "example": 10
                },
                "organized_at": {
                    "type": "string",
                    "example": "2026-04-12T10:05:00Z"
                },
                "progress": {
                    "type": "number",
                    "example": 0.9
                },
                "reverted_at": {
                    "type": "string",
                    "example": "2026-04-12T11:00:00Z"
                },
                "reverted_count": {
                    "type": "integer",
                    "example": 7
                },
                "started_at": {
                    "type": "string",
                    "example": "2026-04-12T10:00:00Z"
                },
                "status": {
                    "allOf": [
                        {
                            "$ref": "#/definitions/github_com_javinizer_javinizer-go_internal_models.JobStatus"
                        }
                    ],
                    "example": "organized"
                },
                "total_files": {
                    "type": "integer",
                    "example": 10
                }
            }
        },
...
        "github_com_javinizer_javinizer-go_internal_api_contracts.OperationItem": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string",
                    "example": "2026-04-12T10:05:00Z"
                },
                "id": {
                    "type": "integer",
                    "example": 1
                },
                "in_place_renamed": {
                    "type": "boolean",
                    "example": false
                },
                "movie_id": {
                    "type": "string",
                    "example": "ABC-123"
                },
                "new_path": {
                    "type": "string",
                    "example": "/dest/ABC-123 [Studio]/ABC-123.mp4"
                },
                "operation_type": {
                    "allOf": [
                        {
                            "$ref": "#/definitions/github_com_javinizer_javinizer-go_internal_models.OperationTypeEnum"
                        }
                    ],
                    "example": "move"
                },
                "original_path": {
                    "type": "string",
                    "example": "/source/ABC-123.mp4"
                },
                "revert_status": {
                    "allOf": [
                        {
                            "$ref": "#/definitions/github_com_javinizer_javinizer-go_internal_models.RevertStatusEnum"
                        }
                    ],
                    "example": "pending"
                },
                "reverted_at": {
                    "type": "string",
                    "example": "2026-04-12T11:00:00Z"
                }
            }
        },
        "github_com_javinizer_javinizer-go_internal_api_contracts.OperationListResponse": {
            "type": "object",
            "properties": {
                "job_id": {
                    "type": "string",
                    "example": "550e8400-e29b-41d4-a716-446655440000"
                },
                "job_status": {
                    "allOf": [
                        {
                            "$ref": "#/definitions/github_com_javinizer_javinizer-go_internal_models.JobStatus"
                        }
                    ],
                    "example": "organized"
                },
                "operations": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/github_com_javinizer_javinizer-go_internal_api_contracts.OperationItem"
                    }
                },
                "total": {
                    "type": "integer",
                    "example": 10
                }
            }
        },
...
        "github_com_javinizer_javinizer-go_internal_models.HistoryOperation": {
            "type": "string",
            "enum": [
                "scrape",
                "organize",
                "download",
                "nfo"
            ],
            "x-enum-varnames": [
                "HistoryOpScrape",
                "HistoryOpOrganize",
                "HistoryOpDownload",
                "HistoryOpNFO"
            ]
        },
        "github_com_javinizer_javinizer-go_internal_models.HistoryStatus": {
            "type": "string",
            "enum": [
                "success",
                "failed",
                "reverted"
            ],
            "x-enum-varnames": [
                "HistoryStatusSuccess",
                "HistoryStatusFailed",
                "HistoryStatusReverted"
            ]
        },
        "github_com_javinizer_javinizer-go_internal_models.JobStatus": {
            "type": "string",
            "enum": [
                "pending",
                "running",
                "completed",
                "failed",
                "cancelled",
                "organized",
                "reverted"
            ],
            "x-enum-varnames": [
                "JobStatusPending",
                "JobStatusRunning",
                "JobStatusCompleted",
                "JobStatusFailed",
                "JobStatusCancelled",
                "JobStatusOrganized",
                "JobStatusReverted"
            ]
        },
...
        "github_com_javinizer_javinizer-go_internal_models.OperationTypeEnum": {
            "type": "string",
            "enum": [
                "move",
                "copy",
                "hardlink",
                "symlink",
                "update"
            ],
            "x-enum-comments": {
                "OperationTypeUpdate": "update-mode organize (NFO overwrite, no file move) per HIST-05"
            },
            "x-enum-descriptions": [
                "",
                "",
                "",
                "",
                "update-mode organize (NFO overwrite, no file move) per HIST-05"
            ],
            "x-enum-varnames": [
                "OperationTypeMove",
                "OperationTypeCopy",
                "OperationTypeHardlink",
                "OperationTypeSymlink",
                "OperationTypeUpdate"
            ]
        },
...
        "github_com_javinizer_javinizer-go_internal_models.RevertStatusEnum": {
            "type": "string",
            "enum": [
                "applied",
                "reverted",
                "failed"
            ],
            "x-enum-comments": {
                "RevertStatusApplied": "Renamed from \"pending\" — D-01"
            },
            "x-enum-descriptions": [
                "Renamed from \"pending\" — D-01",
                "",
                ""
            ],
            "x-enum-varnames": [
                "RevertStatusApplied",
                "RevertStatusReverted",
                "RevertStatusFailed"
            ]
        },
...
        "internal_api_history.DeleteHistoryBulkResponse": {
            "type": "object",
            "properties": {
                "deleted": {
                    "type": "integer"
                }
            }
        },
        "internal_api_history.HistoryListResponse": {
            "type": "object",
            "properties": {
                "limit": {
                    "type": "integer"
                },
                "offset": {
                    "type": "integer"
                },
                "records": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/internal_api_history.HistoryRecord"
                    }
                },
                "total": {
                    "type": "integer"
                }
            }
        },
        "internal_api_history.HistoryRecord": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string"
                },
                "dry_run": {
                    "type": "boolean"
                },
                "error_message": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "metadata": {
                    "type": "string"
                },
                "movie_id": {
                    "type": "string"
                },
                "new_path": {
                    "type": "string"
                },
                "operation": {
                    "$ref": "#/definitions/github_com_javinizer_javinizer-go_internal_models.HistoryOperation"
                },
                "original_path": {
                    "type": "string"
                },
                "status": {
                    "$ref": "#/definitions/github_com_javinizer_javinizer-go_internal_models.HistoryStatus"
                }
            }
        },
...
    }
}